+++ log/tracing_logger_test.go
@@ -14,8 +14,7 @@ import (
 func TestTracingLogger(t *testing.T) {
 	var buf bytes.Buffer
 
-	var logger log.Logger
-	logger = log.NewTMJSONLogger(&buf)
+	logger := log.NewTMJSONLogger(&buf)
 
 	logger1 := log.NewTracingLogger(logger)
 	err1 := errors.New("Courage is grace under pressure.")